deduplicate json input formats
